+++ core/engine/src/value/tests.rs
@@ -569,7 +569,7 @@ fn to_length() {
         assert_eq!(JsValue::new(f64::NEG_INFINITY).to_length(ctx).unwrap(), 0);
         assert_eq!(
             JsValue::new(f64::INFINITY).to_length(ctx).unwrap(),
-            Number::MAX_SAFE_INTEGER as u64
+            Number::MAX_SAFE_INTEGER as usize
         );
         assert_eq!(JsValue::new(0.0).to_length(ctx).unwrap(), 0);
         assert_eq!(JsValue::new(-0.0).to_length(ctx).unwrap(), 0);
